Make the automation more user friendly.
When an user unchecks an "auto" checkbox and waits for the slots to be filled up. Ex: There is a potato in the inventory waiting to be converted to seeds. When they check the checkbox the automation does nothing. You have to manually convert the seed the first time for it to kick back into action.